requirements.txt
django-ckeditor
pillow
pytils
python-dateutil
django-modeltranslation
django-grappelli
